+++ Assets/Script/CleanPlaceData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5712f8ee2acfd514a970076a08775691
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
