+++ Assets/Script/CleanPlaceData/SEDataTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c70f915a8741a84386ed0bb44ef0167
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
