+++ Assets/Script/CleanPlaceData/TimeCalucurator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5814446c3b8e0e047bb7c924b4932207
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
