+++ Assets/Script/CleanPlaceData/TimeConvert.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9fd482b54b2aaba468e74b1f3eab3c02
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
