+++ Assets/Script/Main_UI.cs
@@ -140,13 +140,14 @@ public class Main_UI : MainBase,
     //+ボタンから追加ウィンドウを表示
     public void ChangeAddPlaceMode()
     {
-        SetInputData("i");
+        AddInputData("i");
+
         Enter();
     }
     //キャンセル押したときに追加ウィンドウを閉じる
     public void ChangeDisplayMode()
     {
-        SetInputData("display");
+        AddInputData("display");
         Enter();
     }
     public void OpenPlaceDataMode(int n)
@@ -156,23 +157,23 @@ public class Main_UI : MainBase,
             n = nowTargetIndex; //今の開いているplacedataの場所を代入
         }
         SetTargetIndex(n);
-        SetInputData("placeData");
+        AddInputData("placeData");
         Enter();
     }
     //データの追加と表示
     public void AddPlaceData()
     {
-        SetInputData(addPlaceInputField.textComponent.text);
+        AddInputData(addPlaceInputField.textComponent.text);
         Enter();
     }
     public void RenamePlaceData()
     {
-        SetInputData(renamePlaceInputField.textComponent.text);
+        AddInputData(renamePlaceInputField.textComponent.text);
         Enter();
     }
     public void ChangeResetMode()
     {
-        SetInputData("reset");
+        AddInputData("reset");
         Enter();
     }
     /// <summary>
@@ -180,23 +181,23 @@ public class Main_UI : MainBase,
     /// </summary>
     public void RemovePlaceData()
     {
-        SetInputData("remove");
+        AddInputData("remove");
         Enter();
 
-        StartCoroutine(WaitFrame(1, () => SetInputData(nowTargetIndex.ToString())));
+        StartCoroutine(WaitFrame(1, () => AddInputData(nowTargetIndex.ToString())));
         StartCoroutine(WaitFrame(1, () => Enter()));
     }
 
 
     public void ChengeSetIntervalMode()
     {
-        SetInputData("interval");
+        AddInputData("interval");
         Enter();
     }
 
     public void SetIntervalData()
     {
-        SetInputData(setIntervalDataInputField.textComponent.text);
+        AddInputData(setIntervalDataInputField.textComponent.text);
         Enter();
     }
     /// <summary>
@@ -204,7 +205,7 @@ public class Main_UI : MainBase,
     /// </summary>
     public void ChangeChangeMode()
     {
-        SetInputData("change");
+        AddInputData("change");
         Enter();
     }
     /// <summary>
@@ -212,23 +213,23 @@ public class Main_UI : MainBase,
     /// </summary>
     public void ChangeRemoveCheckMode()
     {
-        SetInputData("removecheck");
+        AddInputData("removecheck");
         Enter();
     }
     public void ChangePlaceDataMode()
     {
-        SetInputData("dataUpdateToPlaceData");
+        AddInputData("dataUpdateToPlaceData");
         Enter();
     }
     public void ChangeRenameMode()
     {
-        SetInputData("rename");
+        AddInputData("rename");
         Enter();
     }
 
     public void OpenOption()
     {
-        SetInputData("option");
+        AddInputData("option");
         Enter();
     }
 
@@ -291,13 +292,13 @@ public class Main_UI : MainBase,
 
     public void SetIntervalDataDropdownDay()
     {
-        SetInputData(setIntervalDataDropdownDay.captionText.text);
+        AddInputData(setIntervalDataDropdownDay.captionText.text);
         Enter();
     }
 
     public void SetIntervalDataDropdownNumber()
     {
-        SetInputData(setIntervalDataDropdownNumber.captionText.text);
+        AddInputData(setIntervalDataDropdownNumber.captionText.text);
         Enter();
     }
 
@@ -308,18 +309,25 @@ public class Main_UI : MainBase,
     //DayとNumberのデータを受け取るときのインターフェース
     public void RecieveDayAndNumberAction(string Day, int number)
     {
+        if (currentMode == CurrentMode.SETINTERVALMODE)
+        {
+            AddInputData(Day);
+            Enter();
+            StartCoroutine(WaitFrame(1, () => AddInputData(number.ToString())));
+            StartCoroutine(WaitFrame(1, () => Enter())); 
+        }
+        else if (currentMode == CurrentMode.ADDPLACEMODE)
+        {
 
-        SetInputData(Day);
-        Enter();
-
-        StartCoroutine(WaitFrame(1, () => SetInputData(number.ToString())));
-        StartCoroutine(WaitFrame(1, () => Enter())); 
+            AddInputData(Day);
+            AddInputData(number.ToString());
+        }
         
     }
 
     public void RecivePushTimeNumber(string num)
     {
-        SetInputData(num);
+        AddInputData(num);
         Enter();
     }
     #endregion
