+++ Assets/cleanPlaceData.json
@@ -1,1 +1,1 @@
-{"placeDataList":[{"place":"ddd","lastUpdateTime_forSave":{"year":2019,"month":12,"day":12,"hour":17,"minute":5,"second":19},"cleanInterval_forSave":{"year":0,"month":0,"day":3,"hour":0,"minute":0,"second":0},"leftDate":-1},{"place":"a","lastUpdateTime_forSave":{"year":2019,"month":12,"day":12,"hour":15,"minute":59,"second":24},"cleanInterval_forSave":{"year":0,"month":0,"day":0,"hour":0,"minute":0,"second":0},"leftDate":0},{"place":"b","lastUpdateTime_forSave":{"year":2019,"month":12,"day":12,"hour":15,"minute":59,"second":31},"cleanInterval_forSave":{"year":0,"month":0,"day":1,"hour":0,"minute":0,"second":0},"leftDate":0},{"place":"a","lastUpdateTime_forSave":{"year":2019,"month":12,"day":12,"hour":16,"minute":10,"second":46},"cleanInterval_forSave":{"year":0,"month":0,"day":3,"hour":0,"minute":0,"second":0},"leftDate":2},{"place":"test","lastUpdateTime_forSave":{"year":2019,"month":12,"day":12,"hour":17,"minute":4,"second":3},"cleanInterval_forSave":{"year":0,"month":0,"day":5,"hour":0,"minute":0,"second":0},"leftDate":4}],"pushTiming":13}
+{"placeDataList":[],"pushTiming":3}
